CORE-8792 Move creation of GroupDetailsDialog to presenter for editing a Group
