+++ de-lib/src/main/java/org/iplantc/de/collaborators/client/ManageCollaboratorsView.java
@@ -4,6 +4,7 @@ import org.iplantc.de.client.models.collaborators.Collaborator;
 import org.iplantc.de.client.models.groups.Group;
 import org.iplantc.de.collaborators.client.events.AddGroupSelected;
 import org.iplantc.de.collaborators.client.events.DeleteGroupSelected;
+import org.iplantc.de.collaborators.client.events.GroupNameSelected;
 import org.iplantc.de.collaborators.client.events.RemoveCollaboratorSelected;
 
 import com.google.gwt.resources.client.ImageResource;
@@ -23,7 +24,8 @@ import java.util.List;
 public interface ManageCollaboratorsView extends IsWidget,
                                                  RemoveCollaboratorSelected.HasRemoveCollaboratorSelectedHandlers,
                                                  DeleteGroupSelected.HasDeleteGroupSelectedHandlers,
-                                                 AddGroupSelected.HasAddGroupSelectedHandlers {
+                                                 AddGroupSelected.HasAddGroupSelectedHandlers,
+                                                 GroupNameSelected.HasGroupNameSelectedHandlers {
 
     /**
      * Appearance related items for the ManageCollaboratorsView
